# Conflicts:
#	src/main/java/com/elster/jupiter/pki/rest/impl/MessageSeeds.java
#	src/main/java/com/elster/jupiter/pki/rest/impl/TrustStoreResource.java

+++ copl/com.elster.jupiter.pki.rest/src/main/java/com/elster/jupiter/pki/rest/impl/CertificateWrapperResource.java
@@ -24,7 +24,7 @@ import com.elster.jupiter.rest.util.JsonQueryParameters;
 import com.elster.jupiter.rest.util.PagedInfoList;
 import com.elster.jupiter.rest.util.Transactional;
 import com.elster.jupiter.util.Checks;
-
+import com.elster.jupiter.util.conditions.Where;
 import org.bouncycastle.asn1.ASN1ObjectIdentifier;
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x500.X500NameBuilder;
@@ -206,6 +206,9 @@ public class CertificateWrapperResource {
     public Response removeCertificate(@PathParam("id") long certificateId) {
         CertificateWrapper certificateWrapper = securityManagementService.findCertificateWrapper(certificateId)
                 .orElseThrow(exceptionFactory.newExceptionSupplier(MessageSeeds.NO_SUCH_CERTIFICATE));
+        if (!securityManagementService.getDirectoryCertificateUsagesQuery().select(Where.where("certificate").isEqualTo(certificateWrapper)).isEmpty()) {
+            throw exceptionFactory.newException(MessageSeeds.CERTIFICATE_USED_BY_DIRECTORY, certificateId);
+        }
         certificateWrapper.delete();
         return Response.status(Response.Status.OK).build();
     }
